move ComitConf to separate file from test_lib
