+++ api_tests/e2e/rfc003/btc_eth-erc20/test.js
@@ -1,6 +1,7 @@
 const chai = require("chai");
 chai.use(require("chai-http"));
 const test_lib = require("../../../test_lib.js");
+const comit_node = require("../../../comit_node_conf.js");
 const should = chai.should();
 const ethutil = require("ethereumjs-util");
 
@@ -14,8 +15,8 @@ const alice_initial_eth = "0.1";
 const bob_initial_eth = "5";
 const bob_initial_erc20 = web3.utils.toWei("10000", "ether");
 
-const alice = test_lib.comit_conf("alice", {});
-const bob = test_lib.comit_conf("bob", {});
+const alice = comit_node.create("alice", {});
+const bob = comit_node.create("bob", {});
 
 const alice_final_address = "0x00a329c0648769a73afac7f9381e08fb43dbea72";
 const bob_final_address =
